requirements.txt
python-telegram-bot==20.7
Flask
requests
psycopg2-binary
SQLAlchemy
Flask-Login